+++ CHANGELOG.md
@@ -6,6 +6,7 @@ New:
 
 - minimization via bisimulation
 - multipointed models, action models, structures and events
+- added Cheryl's Birthday example
 
 Changed: